requirements/base.txt
asgiref==3.2.10
Django==3.1
pytz==2020.1
sqlparse==0.3.1
django-environ==0.4.5
psycopg2-binary==2.8.5
djangorestframework==3.11.1
